Move accountshandler into two different files
