+++ api/include/DBHandlers/SettingsHandler.php
@@ -32,22 +32,4 @@ class SettingsHandler extends UserKnownHandler {
     }
 }
 
-class UserHandler extends UserNotKnownHandler {
-    function __construct() {
-        parent::__construct();
-    }
-
-    public function createNewUser() {
-        return [
-            "statusCode" => 200
-        ];
-    }
-
-    public function getUser($email, $password) {
-        return [
-            "statusCode" => 200
-        ];
-    }
-}
-
 ?>
